separated cli into modules
